requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.4
distro==1.9.0
Django==6.0.1
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.20.3
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.62.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.3
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.2
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
packaging==26.0
pandas==2.2.3
pillow==12.1.0
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.7
pyparsing==3.3.2
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
reportlab==4.0.4
requests==2.32.5
rsa==4.9.1
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.10.2
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
tenacity==9.1.3
threadpoolctl==3.6.0
tifffile==2026.1.28
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
